Change views to float type and restore image upload

+++ recipesAPI/profiles/serializers.py
@@ -3,5 +3,5 @@ from rest_framework import serializers
 
 class UserProfileSerializer(serializers.Serializer):
     name = serializers.CharField()
-    rating = serializers.IntegerField(allow_null=True)
+    rating = serializers.FloatField(allow_null=True)
     views = serializers.IntegerField(allow_null=True)
